feat(gym-app): implement gym-app fundementals, structure, and basic application.
This commit implements a range of files, but most importantly implements the ability to generate QR codes from a Roblox experience, this is good because I don't want to use my pin to enter my Gym, and i'd much rather have a Roblox game do it for me. Also +10 pts on how cool this project is.